Scheduler creating class for nav graph composable and moved things there
